requirements.txt
Flask
flask-cors
tensorflow==2.21.0
numpy
Pillow
Werkzeug